Table loads binary data; needs to be tested.

+++ include/tr_utils.hpp
@@ -23,7 +23,7 @@ T SwapEndianness(T var) {
 
   source.var = var;
 
-  for (int k = 0; k < sizeof(T); k++) {
+  for (size_t k = 0; k < sizeof(T); k++) {
     dest.u8[k] = source.u8[sizeof(T) - k - 1];
   }
